[level 0] Title: 대문자와 소문자, Time: 17.09 ms, Memory: 74 MB -BaekjoonHub

+++ 프로그래머스/0/120893. 대문자와 소문자/대문자와 소문자.java
@@ -4,10 +4,10 @@ class Solution {
         
         for(int i = 0; i < my_string.length(); i++) {
             char alphabet = my_string.charAt(i);
-            if((int) alphabet >= 97)
-                answer += (char) ((int) alphabet - 32);
+            if(alphabet >= 97)
+                answer += (char)(alphabet - 32);
             else
-                answer += (char) ((int) alphabet + 32);
+                answer += (char)(alphabet + 32);
         }
         
         return answer;
